CloudFunction/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
db-dtypes
pandas
evidently